Restructure src into a base layer, named concerns and advanced/
Pure code motion — no behavior change. 324 exports before, 324 after; the
2129 tests pass and every spec golden holds except bundleSize, which moved
+14 bytes (22027 -> 22041) because the module boundaries shifted esbuild's
output, not because anything shipped that didn't before.

base.ts absorbs types + schema + tags + flags + path, plus the `Builder` and
`Encoder` aliases that only ever pointed back at `Val`. Those five were
mutually dependent by nature, so splitting them bought import churn and two
type-only up-edges and nothing else. base.ts now has no outgoing edge at
all, which is the property that makes it safe for everything to reach.

Three modules had stopped describing their contents. refinements.ts held the
public schema factories and the whole compactColumns schema; operations.ts
held metadata, transforms, object modes and defaults alongside operation
compilation; formats.ts was json + string formats + date + list + metadata.
They split along what each name claims: factory builds schemas, modifiers
changes them, refinements checks them, operations compiles them. formats.ts
is gone.

src/advanced/ takes one file per schema nothing else builds on — json,
recursive, compactColumns, uint8Array, date, list. `union` stays in the core
because six modules build on it. `setHas` moves to base: it is tag
bookkeeping and only lived in primitives.ts because its first caller did,
which also severs union's last edge to primitives — the union decoder never
depended on a primitive schema, only on that helper. `defsPath` and
`jsonName` move to base too: the modules that recognise them by name can't
import them from their owner without closing a cycle.

The layering is checked, not asserted: no value cycles, and the single
up-edge (operations -> jsonschema, for two JSON Schema type names) is
type-only.
