+++ CLAUDE.md
@@ -10,7 +10,7 @@ Tiebreaker: shortest *generated* code wins over shortest *library* code (runtime
 
 ## Layout
 
-The implementation lives in `packages/sury/src/*.ts` — plain TypeScript, layered acyclically (types → schema → builder → primitives → parse → union → composites → operations → formats → factory → refinements → jsapi → jsonschema; only type-only imports may point "up"). `src/entry.ts` is the single public entry: it re-exports the curated JS/TS API under its documented names, creates the eager PURE-annotated schema constants, and exposes a minimal ReScript-binding surface (`$res_*`-named exports, only for APIs with no public-JS equivalent — `$res` because ReScript externals reject `~` in names). `scripts/pack.ts` bundles it to the gitignored `src/S.mjs` (`pnpm build:entry`); the checked-in `src/S.d.mts` → `S.d.ts` provides its types, and the publish step additionally emits a CJS `S.js` into the artifact for the require condition. `S.res` is the one ReScript module: public types plus `@module("sury") external` bindings resolved through the package's own "." export, so ReScript and JS share a single runtime instance. Where the ReScript API differs from JS only in argument shape, S.res binds the public JS export and adapts in ReScript (`refine`, `to`, `decoder`, `tuple1/2/3`, `parseOrThrow`, …) rather than adding a `$res_*` export. The former `module B` is flattened to `B_`-prefixed top-level functions (and `Literal.parse` → `Literal_parse`, etc.) so bundlers tree-shake each helper individually; keep new helpers flat for the same reason, and PURE-annotate any top-level call initializer. Prefer `const name = () => {}` arrows over `function` declarations (measurably smaller minified; `noopOperation` and the `this`-based `_var` family are the deliberate exceptions), and inline former ReScript intrinsics (`a | b`, `typeof x`) rather than wrapping them in helpers. `val`/`check`/`bGlobal` runtime field names stay short (`cp`, `hd`, `vc`, …) — property names survive minification, so every character ships.
+The implementation lives in `packages/sury/src/*.ts` — plain TypeScript, layered acyclically (base → builder → primitives → parse → union → composites → factory → modifiers → refinements → operations → advanced/\* → jsapi → jsonschema; only type-only imports may point "up", and `operations → jsonschema` is the one that does). `base.ts` is the bottom: the data model (`Internal`, `Val`, `Check`), the schema object and its prototype, tags, flags and paths, plus the `Builder`/`Encoder` aliases — merged into one module because they are mutually dependent by nature, and it has **no outgoing edge at all**, so anything may reach it. The split by concern is what each name says: `factory.ts` builds schemas, `modifiers.ts` takes a schema and returns a changed one (refine, transform, metadata, object modes, defaults), `refinements.ts` layers checks (and the string formats, which are the same idea with a canned predicate), `operations.ts` compiles a schema into a callable. `src/advanced/` holds one file per schema that nothing else builds on — `json`, `recursive`, `compactColumns`, `uint8Array`, `date`, `list` — so a schema with a decoder of its own is read on its own; `union` is *not* among them (six modules build on it, so it stays in the core). Two identity constants live in base rather than with their schema, because the modules that recognise them by name can't import them without a cycle: `defsPath` (`S.recursive`'s `$ref` prefix) and `jsonName`. `src/entry.ts` is the single public entry: it re-exports the curated JS/TS API under its documented names, creates the eager PURE-annotated schema constants, and exposes a minimal ReScript-binding surface (`$res_*`-named exports, only for APIs with no public-JS equivalent — `$res` because ReScript externals reject `~` in names). `scripts/pack.ts` bundles it to the gitignored `src/S.mjs` (`pnpm build:entry`); the checked-in `src/S.d.mts` → `S.d.ts` provides its types, and the publish step additionally emits a CJS `S.js` into the artifact for the require condition. `S.res` is the one ReScript module: public types plus `@module("sury") external` bindings resolved through the package's own "." export, so ReScript and JS share a single runtime instance. Where the ReScript API differs from JS only in argument shape, S.res binds the public JS export and adapts in ReScript (`refine`, `to`, `decoder`, `tuple1/2/3`, `parseOrThrow`, …) rather than adding a `$res_*` export. The former `module B` is flattened to `B_`-prefixed top-level functions (and `Literal.parse` → `Literal_parse`, etc.) so bundlers tree-shake each helper individually; keep new helpers flat for the same reason, and PURE-annotate any top-level call initializer. Prefer `const name = () => {}` arrows over `function` declarations (measurably smaller minified; `noopOperation` and the `this`-based `_var` family are the deliberate exceptions), and inline former ReScript intrinsics (`a | b`, `typeof x`) rather than wrapping them in helpers. `val`/`check`/`bGlobal` runtime field names stay short (`cp`, `hd`, `vc`, …) — property names survive minification, so every character ships.
 
 ## Comments
