+++ packages/sury/specs/bundleSize.yaml
@@ -1,99 +1,99 @@
 # Minified+gzipped bytes per public export of src/S.mjs, plus `total` for the whole entry.
 # Generated by `pnpm spec check --write` — every row is measured, so never hand-write one.
-total: 22027
+total: 22041
 exports:
-  $res_Metadata_Id_make: 3561
-  $res_Metadata_get: 3556
-  $res_Metadata_set: 3576
-  $res_Option_getOr: 8611
+  $res_Metadata_Id_make: 3562
+  $res_Metadata_get: 3560
+  $res_Metadata_set: 3582
+  $res_Option_getOr: 8612
   $res_Option_getOrWith: 8611
-  $res_assertAsyncOrThrow: 3916
-  $res_nullAsOption: 10563
+  $res_assertAsyncOrThrow: 3917
+  $res_nullAsOption: 10568
   $res_nullAsUnit: 3938
-  $res_nullableAsOption: 7946
-  $res_option: 10542
-  $res_pathConcat: 3548
+  $res_nullableAsOption: 7954
+  $res_option: 10541
+  $res_pathConcat: 3549
   $res_pathFromArray: 3639
   $res_pathFromLocation: 3595
-  $res_pathToArray: 3548
-  $res_schema: 10766
-  $res_setExnId: 3558
-  $res_transform: 4070
-  $res_unit: 3898
-  Error: 3553
+  $res_pathToArray: 3549
+  $res_schema: 10772
+  $res_setExnId: 3560
+  $res_transform: 4073
+  $res_unit: 3901
+  Error: 3554
   any: 3551
-  array: 10542
-  assert: 3952
-  asyncDecoder: 3560
-  asyncDecoderAssert: 4100
-  asyncEncoder: 3910
-  asyncParser: 3560
+  array: 10539
+  assert: 3954
+  asyncDecoder: 3563
+  asyncDecoderAssert: 4103
+  asyncEncoder: 3912
+  asyncParser: 3564
   bigint: 3862
-  bool: 3856
-  boolean: 3856
-  brand: 3561
-  compactColumns: 11406
+  bool: 3857
+  boolean: 3857
+  brand: 3565
+  compactColumns: 11411
   cuid: 3967
-  date: 3899
-  decoder: 3548
-  deepStrict: 3664
-  deepStrip: 3662
-  dict: 10545
-  email: 4012
-  enableStandardJSONSchema: 5322
-  encoder: 3908
-  enum: 10766
-  extendJSONSchema: 3633
-  float: 3951
-  fromJSONSchema: 16528
-  global: 3597
-  instance: 3709
-  int: 3959
-  int32: 3959
-  is: 3981
-  isAsync: 3610
-  isoDateTime: 4005
-  json: 11925
-  jsonString: 12390
-  jsonStringWithSpace: 12400
+  date: 3897
+  decoder: 3552
+  deepStrict: 3668
+  deepStrip: 3666
+  dict: 10546
+  email: 4014
+  enableStandardJSONSchema: 5324
+  encoder: 3910
+  enum: 10770
+  extendJSONSchema: 3635
+  float: 3955
+  fromJSONSchema: 16512
+  global: 3600
+  instance: 3715
+  int: 3962
+  int32: 3962
+  is: 3982
+  isAsync: 3611
+  isoDateTime: 4007
+  json: 11927
+  jsonString: 12393
+  jsonStringWithSpace: 12402
   length: 3950
-  list: 11085
-  literal: 10757
-  max: 4021
+  list: 11092
+  literal: 10762
+  max: 4023
   merge: 10664
-  meta: 4039
-  min: 4019
-  nan: 3906
+  meta: 4045
+  min: 4023
+  nan: 3907
   never: 3640
-  noValidation: 3562
-  nullable: 8686
-  nullish: 7925
-  number: 3951
-  object: 12813
-  optional: 8661
-  parser: 3558
-  pattern: 3754
-  port: 4043
-  record: 10545
-  recursive: 4196
-  refine: 3757
-  reverse: 3897
-  safe: 3591
-  safeAsync: 3606
-  schema: 10757
-  shape: 11987
-  strict: 3664
-  string: 3851
-  strip: 3661
-  symbol: 3777
+  noValidation: 3565
+  nullable: 8685
+  nullish: 7932
+  number: 3955
+  object: 12821
+  optional: 8655
+  parser: 3563
+  pattern: 3753
+  port: 4045
+  record: 10546
+  recursive: 4197
+  refine: 3754
+  reverse: 3898
+  safe: 3595
+  safeAsync: 3611
+  schema: 10762
+  shape: 11988
+  strict: 3667
+  string: 3854
+  strip: 3664
+  symbol: 3782
   to: 3798
   toExpression: 3547
-  toJSONSchema: 5290
-  trim: 4094
-  tuple: 12115
+  toJSONSchema: 5291
+  trim: 4098
+  tuple: 12124
   uint8Array: 4025
-  union: 10759
+  union: 10765
   unknown: 3551
-  url: 3963
+  url: 3964
   uuid: 3984
-  void: 3907
+  void: 3909
