+++ packages/sury/src/union.ts
@@ -1,7 +1,46 @@
-import { setHas } from "./primitives";
-import { baseSchema, getOrRethrow, panic, unknown, updateOutput } from "./schema";
-import { getOutputSchema, nestedLoc, never_, parse, typeCheckCond } from "./parse";
 import {
+  baseSchema,
+  type Builder,
+  type Check,
+  type Encoder,
+  flagUnsafeHas,
+  getOrRethrow,
+  immutableEmptyArray,
+  immutableEmptyObject,
+  type Internal,
+  isLiteral,
+  neverTag,
+  nullTag,
+  numberTag,
+  objectTag,
+  panic,
+  setHas,
+  type SuryErrorRecord,
+  type Tag,
+  tagFlagArray,
+  tagFlagFunction,
+  tagFlagInstance,
+  tagFlagNaN,
+  tagFlagNever,
+  tagFlagNull,
+  tagFlagObject,
+  tagFlagRef,
+  tagFlags,
+  tagFlagUndefined,
+  tagFlagUnion,
+  tagFlagUnknown,
+  toExpression,
+  U,
+  undefinedTag,
+  unionTag,
+  unknown,
+  updateOutput,
+  type Val,
+  valFlagAsync,
+} from "./base";
+import {
+  _notVar,
+  _var,
   B_embed,
   B_failWithArg,
   B_invalidOperation,
@@ -13,47 +52,10 @@ import {
   B_refine,
   B_scope,
   B_throw,
-  Builder,
-  Encoder,
-  HoistCond,
-  _notVar,
-  _var,
   failInvalidType,
+  type HoistCond,
 } from "./builder";
-import {
-  Check,
-  Internal,
-  SuryErrorRecord,
-  U,
-  Val,
-  immutableEmptyArray,
-  immutableEmptyObject,
-  isLiteral,
-  toExpression,
-} from "./types";
-import { flagUnsafeHas, valFlagAsync } from "./flags";
-import {
-  Tag,
-  neverTag,
-  nullTag,
-  numberTag,
-  objectTag,
-  tagFlagArray,
-  tagFlagFunction,
-  tagFlagInstance,
-  tagFlagNaN,
-  tagFlagNever,
-  tagFlagNull,
-  tagFlagNumber,
-  tagFlagObject,
-  tagFlagRef,
-  tagFlagUndefined,
-  tagFlagUnion,
-  tagFlagUnknown,
-  tagFlags,
-  undefinedTag,
-  unionTag,
-} from "./tags";
+import { getOutputSchema, nestedLoc, never_, parse, typeCheckCond } from "./parse";
 
 // Every tag bit set — the acceptance mask of a case that narrows nothing, and of
 // a source that can hold any runtime type.
